+++ nodes/__init__.py
@@ -1,6 +1,6 @@
 """ Node classes used by the Notification Node Server. """
 
-NSVERSION = "3.6.26"
+NSVERSION = "3.6.27"
 
 from .Notify import Notify
 from .Pushover import Pushover